+++ native/cocos/platform/interfaces/modules/canvas/ICanvasRenderingContext2D.h
@@ -80,11 +80,11 @@ public:
         virtual void updateFont(const ccstd::string &fontName, float fontSize, bool bold, bool italic, bool oblique, bool smallCaps) = 0;
         virtual void setTextAlign(TextAlign align) = 0;
         virtual void setTextBaseline(TextBaseline baseline) = 0;
-        virtual void setFillStyle(uint8_t r, uint8_t g, uint8_t b, float a) = 0;
-        virtual void setStrokeStyle(uint8_t r, uint8_t g, uint8_t b, float a) = 0;
+        virtual void setFillStyle(uint8_t r, uint8_t g, uint8_t b, uint8_t a) = 0;
+        virtual void setStrokeStyle(uint8_t r, uint8_t g, uint8_t b, uint8_t a) = 0;
         virtual void setLineWidth(float lineWidth) = 0;
         virtual void setShadowBlur(float blur) = 0;
-        virtual void setShadowColor(uint8_t r, uint8_t g, uint8_t b, float a) = 0;
+        virtual void setShadowColor(uint8_t r, uint8_t g, uint8_t b, uint8_t a) = 0;
         virtual void setShadowOffsetX(float offsetX) = 0;
         virtual void setShadowOffsetY(float offsetY) = 0;
         virtual const cc::Data &getDataRef() const = 0;
